BIG CHANGE!
For Raster,  moved groundGeom and acquistionDate to the Metadata object.  This was
done mainly for performance reasons to cut down on the number of joins that has to
be done for querying.   It probably makes sense for them to be in that object anyway.

The same was done for Video's groundGeom, startDate and endDate.



git-svn-id: http://svn.osgeo.org/ossim/omar@16672 b630af17-092b-0410-afba-a002956477a4