init(): I added a params check for rotateAngle so the user can control the rotation via the REST interface.
changeToSignleLayer(): I added a POST call to determine the center lat/lon of the map when the user is going from image space to ground space. That center point is then passed to the ground space REST so the user wont lose their place on the map when going from image space to ground space.

shareImage(): I added a rotation parameter so the user captures the image's rotation when sharing.


git-svn-id: http://svn.osgeo.org/ossim/omar@20967 b630af17-092b-0410-afba-a002956477a4